new commit

+++ src/app/app.component.ts
@@ -1,7 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { Observable } from 'rxjs';
+import { User } from './API Services/for Booking/client';
 import { ExperTexhService } from './API Services/for Booking/exper-texh.service';
 import { ReportingService } from './API Services/for User/reporting.service';
+import { UserData } from './Staff/setup/setup.component';
 
 @Component({
   selector: 'app-root',
@@ -13,6 +16,7 @@ export class AppComponent implements OnInit {
   title = 'Screens';
   showBadges = true;
   RoleID = null;
+  User;
   
   toggleSidebar()
   {
@@ -30,8 +34,43 @@ export class AppComponent implements OnInit {
     }
 
     this.api.getBadgeCount();
-    
+    this.api.getProfile().subscribe(data => {
+      //console.log("User Details",data.ContactNo)
+      //this.User = data;
+      this.mapValues(data)
+    })
   }
+
+mapValues(data:User)
+{
+  if(this.api.RoleID == "1" )
+  {
+    this.User =
+    {
+      Name: data.Clients[0].Name,
+      Surname: data.Clients[0].Surname,
+      Email: data.Clients[0].Email,
+    }
+  }
+  else if(this.api.RoleID == "2" )
+  {
+    this.User =
+    {
+      Name: data.Admins[0].Name,
+      Surname: data.Admins[0].Surname,
+      Email: data.Admins[0].Email,
+    }
+  }
+  else if(this.api.RoleID == "3" )
+  {
+    this.User =
+    {
+      Name: data.Employees[0].Name,
+      Surname: data.Employees[0].Surname,
+      Email: data.Employees[0].Email,
+    }
+  }
+}
 
   login()
   {
@@ -48,16 +87,22 @@ export class AppComponent implements OnInit {
 
   logout()
   {
-    sessionStorage.clear();
+    this.api.logout().subscribe(res =>
+      {
+        if(res == "success")
+        {
+        sessionStorage.clear();
 
-    this.RoleID = null;
-    alert("Logged out successfully. Re-directing to homepage");
-
-    this.router.navigateByUrl('/home')
-    .then(() => {
-      window.location.reload();
-    });
-
+        this.RoleID = null;
+    
+        this.router.navigateByUrl('/home')
+        .then(() => {
+          window.location.reload();
+        });
+        }
+    
+      })
+    
   }
   
 }
